Fixes generated go code for time.Time fields resulting in parsing errors
This fixes Go code generated with time.Time which would cause parsing errors on any datetime values not
compliant with RFC-3339. The [xsd:datetime spec](https://www.w3.org/TR/xmlschema-2/#dateTime) states
that the format is inspired by ISO-8601 which is [not the same](https://ijmacd.github.io/rfc3339-iso8601/) as RFC-3339.
As a result, prior to this change, the generated code was likely to result in parsing errors on those `time.Time` fields.

This changes the mapping of the datetime and related types to `string` for Go. This removes the potential for
parsing errors with the tradeoff that users would now have to do parsing of the `time.Time` values in post-processing.
This seems like the better tradeoff given that users would potentially have data on which time formats their apps
actually need to support.

Fixes #13.

Signed-off-by: Alex Normand <alexandre.normand@gmail.com>