Fix directory creation and filename extension handling
Fixes a bug where xgen assumes that the output path provided
by the user is a directory. This removes the call to
PrepareOutputDir from xgen.go, but this should be fine because
PrepareOutputDir is also called, correctly, in parser.go.

Fixes a bug where the language-specific generator code assumes
that the output path does not contain the file extension. This
commit adds a helper function called FileWithExtension that
checks to see if the extension already exists on the filename
and does not add it again if it does.