Change position of opt.CurrentEle = opt.InElement
The line in the outer position of 'if' causes the simpleType element to
be set to CurrentEle every time it is parsed. However, simpleType is not
always the parent of all nested elements. If I have the following
relationship: <element> -> <simpleType> -> [some element], then
CurrentEle should be the element, not simpleType.